update request route
