+++ client/src/pages/Login.js
@@ -60,7 +60,6 @@ class Login extends Component {
         const decoded = jwt_decode(token);
         localStorage.setItem("jwtToken", token);
         localStorage.setItem("userId", decoded.id);
-        console.log("successfully logged in ");
         this.props.history.push("/profile");
       })
       .catch(err => {
