+++ server/package.json
@@ -31,8 +31,8 @@
     "nodemon": "^1.19.3",
     "passport": "^0.4.0",
     "passport-jwt": "^4.0.0",
-    "stripe": "^7.10.0",
     "socket.io": "^2.3.0",
+    "stripe": "^7.10.0",
     "validator": "^11.1.0"
   },
   "devDependencies": {
