--- server/routes/index.js
@@ -1,8 +1,0 @@
-var express = require("express");
-var router = express.Router();
-
-router.get("/welcome", function(req, res, next) {
-  res.status(200).send({ welcomeMessage: "Step 1 (completed)" });
-});
-
-module.exports = router;
